Basic drawing working (#2)
* Basic drawing working

Now when the player moves it colours in the cells underneath it. We
can also specify the cell size using the x and y scales on the cell
prefab object, the number of cells will be adjusted to suit.

If we make the cells too small we'll very quickly hit performance
issues.

Need to allow the player to turn painitn on and off, as well as turn
erasing on and off.

* updating scene config

* Allow player to turn painting/erasing on/off

Also trying to clean up my filthy Java braces. I just want to fit in.